refactor: extract InventoryGrid into standalone component
